feat: wire channelcluster semantics into access layer

+++ internal/access/api/integration_test.go
@@ -59,7 +59,7 @@ func TestAPIServerSendMessageWithRealMessageApp(t *testing.T) {
 
 	var got struct {
 		MessageID  int64  `json:"message_id"`
-		MessageSeq uint32 `json:"message_seq"`
+		MessageSeq uint64 `json:"message_seq"`
 		Reason     uint8  `json:"reason"`
 	}
 	require.NoError(t, json.NewDecoder(resp.Body).Decode(&got))
